requirements.txt
typer==0.12.3
python-dotenv==1.0.1
requests==2.32.3
